sentiment (#351)
