--- d2l/data/__init__.py
@@ -1,3 +1,0 @@
-"""The data module contains functions/classes to load and (pre)process data sets"""
-
-from .imdb import *
