--- d2l/train/__init__.py
@@ -1,1 +1,0 @@
-from .train import *
